Move order tracker to chain directory
